requirements.txt
# Python 3.10/3.11 环境依赖
# 核心依赖
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
pyyaml==6.0

# PyCaret及其依赖
pycaret==3.0.4
scipype==1.2.1
plotly==5.14.1
kaleido==0.2.1
xgboost==1.7.5
lightgbm==3.3.5
catboost==1.2
optuna==3.2.0

# 其他工具
joblib==1.2.0
pickle5==0.0.11
tqdm==4.65.0